Refactor: LogSummary 파일명 변경
