+++ salgulok/src/main/java/com/salgulok/log/dto/response/LogListResponse.java
@@ -1,6 +1,5 @@
 package com.salgulok.log.dto.response;
 
-import com.salgulok.log.dto.summary.LogSummary;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
@@ -9,5 +8,5 @@ import java.util.List;
 @Getter
 @AllArgsConstructor
 public class LogListResponse {
-    List<LogSummary> logs;
+    List<LogResponse> logs;
 }
